Bug Testing
Lots of new variables being used for bug-testing and variable analysis, do not mind them, most of them are labelled.
The game will pause after a first execution too. Remove a call to the pauseGameExec() function to stop this pausing so that the game moves as usual. 
Ball will go to the top of the screen if it hits the top half of the paddle.